+++ bitalarm-impl/src/test/java/com/dwarfeng/bitalarm/impl/service/CurrentAlarmMaintainServiceImplTest.java
@@ -4,7 +4,7 @@ import com.dwarfeng.bitalarm.stack.bean.entity.AlarmSetting;
 import com.dwarfeng.bitalarm.stack.bean.entity.CurrentAlarm;
 import com.dwarfeng.bitalarm.stack.service.AlarmSettingMaintainService;
 import com.dwarfeng.bitalarm.stack.service.CurrentAlarmMaintainService;
-import com.dwarfeng.subgrade.stack.exception.ServiceException;
+import org.apache.commons.beanutils.BeanUtils;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -14,6 +14,8 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.util.Date;
+
+import static org.junit.Assert.assertEquals;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(locations = "classpath:spring/application-context*.xml")
@@ -40,6 +42,7 @@ public class CurrentAlarmMaintainServiceImplTest {
         );
         currentAlarm = new CurrentAlarm(
                 null,
+                1L,
                 1,
                 "我是报警信息",
                 (byte) 0,
@@ -54,11 +57,13 @@ public class CurrentAlarmMaintainServiceImplTest {
     }
 
     @Test
-    public void test() throws ServiceException {
+    public void test() throws Exception {
         try {
             parentAlarmSetting.setKey(alarmSettingMaintainService.insert(parentAlarmSetting));
             currentAlarm.setKey(parentAlarmSetting.getKey());
             currentAlarmMaintainService.insert(currentAlarm);
+            CurrentAlarm currentAlarm1 = currentAlarmMaintainService.get(this.currentAlarm.getKey());
+            assertEquals(BeanUtils.describe(this.currentAlarm), BeanUtils.describe(currentAlarm1));
         } finally {
             currentAlarmMaintainService.deleteIfExists(currentAlarm.getKey());
             alarmSettingMaintainService.deleteIfExists(parentAlarmSetting.getKey());