+++ bitalarm-sdk/src/main/java/com/dwarfeng/bitalarm/sdk/bean/entity/FastJsonCurrentAlarm.java
@@ -16,7 +16,7 @@ import java.util.Objects;
  */
 public class FastJsonCurrentAlarm implements Bean {
 
-    private static final long serialVersionUID = -6426065913134830677L;
+    private static final long serialVersionUID = 5802461159132091693L;
 
     public static FastJsonCurrentAlarm of(CurrentAlarm currentAlarm) {
         if (Objects.isNull(currentAlarm)) {
@@ -24,6 +24,7 @@ public class FastJsonCurrentAlarm implements Bean {
         } else {
             return new FastJsonCurrentAlarm(
                     FastJsonLongIdKey.of(currentAlarm.getKey()),
+                    currentAlarm.getPointId(),
                     currentAlarm.getIndex(),
                     currentAlarm.getAlarmMessage(),
                     currentAlarm.getAlarmType(),
@@ -35,23 +36,28 @@ public class FastJsonCurrentAlarm implements Bean {
     @JSONField(name = "key", ordinal = 1)
     private FastJsonLongIdKey key;
 
-    @JSONField(name = "index", ordinal = 2)
+    @JSONField(name = "point_id", ordinal = 2)
+    private long pointId;
+
+    @JSONField(name = "index", ordinal = 3)
     private int index;
 
-    @JSONField(name = "alarm_message", ordinal = 3)
+    @JSONField(name = "alarm_message", ordinal = 4)
     private String alarmMessage;
 
-    @JSONField(name = "alarm_type", ordinal = 4)
+    @JSONField(name = "alarm_type", ordinal = 5)
     private byte alarmType;
 
-    @JSONField(name = "happened_date", ordinal = 5)
+    @JSONField(name = "happened_date", ordinal = 6)
     private Date happenedDate;
 
     public FastJsonCurrentAlarm() {
     }
 
-    public FastJsonCurrentAlarm(FastJsonLongIdKey key, int index, String alarmMessage, byte alarmType, Date happenedDate) {
+    public FastJsonCurrentAlarm(
+            FastJsonLongIdKey key, long pointId, int index, String alarmMessage, byte alarmType, Date happenedDate) {
         this.key = key;
+        this.pointId = pointId;
         this.index = index;
         this.alarmMessage = alarmMessage;
         this.alarmType = alarmType;
@@ -64,6 +70,14 @@ public class FastJsonCurrentAlarm implements Bean {
 
     public void setKey(FastJsonLongIdKey key) {
         this.key = key;
+    }
+
+    public long getPointId() {
+        return pointId;
+    }
+
+    public void setPointId(long pointId) {
+        this.pointId = pointId;
     }
 
     public int getIndex() {
@@ -102,6 +116,7 @@ public class FastJsonCurrentAlarm implements Bean {
     public String toString() {
         return "FastJsonCurrentAlarm{" +
                 "key=" + key +
+                ", pointId=" + pointId +
                 ", index=" + index +
                 ", alarmMessage='" + alarmMessage + '\'' +
                 ", alarmType=" + alarmType +