+++ bitalarm-sdk/src/main/java/com/dwarfeng/bitalarm/sdk/bean/entity/JSFixedFastJsonAlarmInfo.java
@@ -1,6 +1,7 @@
 package com.dwarfeng.bitalarm.sdk.bean.entity;
 
 import com.alibaba.fastjson.annotation.JSONField;
+import com.alibaba.fastjson.serializer.ToStringSerializer;
 import com.dwarfeng.bitalarm.stack.bean.entity.AlarmInfo;
 import com.dwarfeng.subgrade.sdk.bean.key.JSFixedFastJsonLongIdKey;
 import com.dwarfeng.subgrade.stack.bean.Bean;
@@ -16,7 +17,7 @@ import java.util.Objects;
  */
 public class JSFixedFastJsonAlarmInfo implements Bean {
 
-    private static final long serialVersionUID = 2576171375273917608L;
+    private static final long serialVersionUID = -7816620407316562535L;
 
     public static JSFixedFastJsonAlarmInfo of(AlarmInfo alarmInfo) {
         if (Objects.isNull(alarmInfo)) {
@@ -24,6 +25,7 @@ public class JSFixedFastJsonAlarmInfo implements Bean {
         } else {
             return new JSFixedFastJsonAlarmInfo(
                     JSFixedFastJsonLongIdKey.of(alarmInfo.getKey()),
+                    alarmInfo.getPointId(),
                     alarmInfo.getIndex(),
                     alarmInfo.getAlarmMessage(),
                     alarmInfo.getAlarmType(),
@@ -36,28 +38,32 @@ public class JSFixedFastJsonAlarmInfo implements Bean {
     @JSONField(name = "key", ordinal = 1)
     private JSFixedFastJsonLongIdKey key;
 
-    @JSONField(name = "index", ordinal = 2)
+    @JSONField(name = "point_id", ordinal = 2, serializeUsing = ToStringSerializer.class)
+    private long pointId;
+
+    @JSONField(name = "index", ordinal = 3)
     private int index;
 
-    @JSONField(name = "alarm_message", ordinal = 3)
+    @JSONField(name = "alarm_message", ordinal = 4)
     private String alarmMessage;
 
-    @JSONField(name = "alarm_type", ordinal = 4)
+    @JSONField(name = "alarm_type", ordinal = 5)
     private byte alarmType;
 
-    @JSONField(name = "happened_date", ordinal = 5)
+    @JSONField(name = "happened_date", ordinal = 6)
     private Date happenedDate;
 
-    @JSONField(name = "alarming", ordinal = 6)
+    @JSONField(name = "alarming", ordinal = 7)
     private boolean alarming;
 
     public JSFixedFastJsonAlarmInfo() {
     }
 
     public JSFixedFastJsonAlarmInfo(
-            JSFixedFastJsonLongIdKey key, int index, String alarmMessage, byte alarmType, Date happenedDate,
-            boolean alarming) {
+            JSFixedFastJsonLongIdKey key, long pointId, int index, String alarmMessage, byte alarmType,
+            Date happenedDate, boolean alarming) {
         this.key = key;
+        this.pointId = pointId;
         this.index = index;
         this.alarmMessage = alarmMessage;
         this.alarmType = alarmType;
@@ -71,6 +77,14 @@ public class JSFixedFastJsonAlarmInfo implements Bean {
 
     public void setKey(JSFixedFastJsonLongIdKey key) {
         this.key = key;
+    }
+
+    public long getPointId() {
+        return pointId;
+    }
+
+    public void setPointId(long pointId) {
+        this.pointId = pointId;
     }
 
     public int getIndex() {
@@ -111,17 +125,5 @@ public class JSFixedFastJsonAlarmInfo implements Bean {
 
     public void setAlarming(boolean alarming) {
         this.alarming = alarming;
-    }
-
-    @Override
-    public String toString() {
-        return "JSFixedFastJsonAlarmInfo{" +
-                "key=" + key +
-                ", index=" + index +
-                ", alarmMessage='" + alarmMessage + '\'' +
-                ", alarmType=" + alarmType +
-                ", happenedDate=" + happenedDate +
-                ", alarming=" + alarming +
-                '}';
     }
 }